MetaReport/requirements.txt
pywin32==227
click==7.1.1
pandas==1.0.1
paramiko==2.7.1
python_docx==0.8.10
future_fstrings
